requirements.txt
torch>=2.0.0
transformers>=4.51.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
wandb>=0.12.0
tqdm>=4.62.0
pytest>=6.2.0
black>=21.5b2
isort>=5.9.0
flake8>=3.9.0 
hydra-core>=1.3.2
omegaconf>=2.3.0